Handle late style updates in old IE and Opera
Older versions of IE and Opera require the height to be set explicitly
as they don't correctly calculate the height: 100%; style in css for the
scrollContainer when using normal browser scrolling.  However, the
containers scrollHeight does not update quickly enough after setting the
height to be used in calculations, so delay it until the next paint.
This also required delaying the scroll event slightly to allow the
update event to complete first.